pyproject.toml
[tool.poetry]
name = "pera-fastapi"
version = "0.1.0"
description = ""
authors = ["Ion Rusu <ionrusu114@ionrusu114.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.0.3"
telethon = "^1.32.0"
sqlalchemy = "^2.0.23"
pymysql = "^1.1.0"
alembic = "^1.12.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
asyncpg = "^0.29.0"
aiomysql = "^0.2.0"
celery = "^5.3.5"
flower = "^2.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
gunicorn = "^21.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.1"
mypy = "^1.6.1"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
